+++ example/lib/main.dart
@@ -15,7 +15,7 @@ class MyApp extends StatelessWidget {
       ),
       home: new MyHomePage(),
       localizationsDelegates: [
-        FlutterI18nDelegate(false, 'en', 'assets/i18n'),
+        FlutterI18nDelegate(path: 'assets/i18n'),
         GlobalMaterialLocalizations.delegate,
         GlobalWidgetsLocalizations.delegate
       ],
@@ -29,12 +29,13 @@ class MyHomePage extends StatefulWidget {
 }
 
 class MyHomeState extends State<MyHomePage> {
-  String currentLang = 'en';
+  Locale currentLang;
   int clicked = 0;
 
   changeLanguage() {
     setState(() {
-      currentLang = currentLang == 'en' ? 'it' : 'en';
+      currentLang =
+          currentLang.languageCode == 'en' ? new Locale('it') : new Locale('en');
     });
   }
 
@@ -46,6 +47,7 @@ class MyHomeState extends State<MyHomePage> {
 
   @override
   Widget build(BuildContext context) {
+    currentLang = FlutterI18n.currentLocale(context);
     return new Scaffold(
       appBar:
           new AppBar(title: new Text(FlutterI18n.translate(context, "title"))),