+++ CHANGELOG.md
@@ -51,3 +51,18 @@ To configure it, use the third optional parameter of `FlutterI18nDelegate`.
 ## [0.5.1]
 
 Print in log when a key is not found.
+
+## [0.5.2]
+
+Fix for bug #10.
+
+## [0.6.0]
+
+Exposed new method that return the locale used by the library:
+
+```sh
+FlutterI18n.currentLocale();
+```
+
+
+***NOTE***: the ***refresh*** method now accept a *Locale* as second parameter, instead of two strings.